Add new delay option.
This option allows multiple file changes to occur before node-dev
restarts itself.  Also add tests and documentation for the option.

With the addition of new tests, increase the testing timeout to 2m.
I don't know enough about TAP to fix it, but it appears the 30s timeout
is applying to the entire file, instead of individual tests.  Therefore
it needed to be increased for this checkin.